feat: extract iMessage into separate FDA-holding helper binary
macOS ties Full Disk Access to the specific binary, so every clawvisor
update forced users to re-grant FDA. This extracts all iMessage database
operations into a small, stable helper binary (clawvisor-imessage-helper)
that holds FDA independently and is installed on demand when the user
first activates iMessage.

The helper has its own protocol version — the main binary only replaces
it (and triggers FDA re-grant) when the protocol actually changes, not
on every update. The helper embeds an Info.plist so it shows as
"Clawvisor iMessage Helper" in System Settings.
