+++ cmd/imessage-helper/Info.plist
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CFBundleIdentifier</key>
+	<string>com.clawvisor.imessage-helper</string>
+	<key>CFBundleName</key>
+	<string>Clawvisor iMessage Helper</string>
+	<key>CFBundleDisplayName</key>
+	<string>Clawvisor iMessage Helper</string>
+</dict>
+</plist>
